Renames src folder.
